Revert "Fix: AccountSetting hasError 수정"
