--- src/pages/UpLoadBoard/components/customQuali/quill-custom.css
@@ -1,95 +1,0 @@
-/* quill-custom.css */
-
-/* 여기에 원하는 스타일을 추가합니다 */
-
-.ql-toolbar {
-  border: 1px solid #e2e8f0 !important;
-  width: 100%;
-  height: 44px;
-  background-color: #ffffff;
-  border-radius: 21.5px;
-  display: flex;
-  position: relative;
-  align-items: center;
-  /* 기타 등등 */
-}
-
-.ql-container {
-  border: none !important;
-  min-height: 480px;
-  height: 100px;
-  resize: none;
-  margin-top: 10px;
-}
-
-.frame-ql-fromats {
-  flex: 2;
-}
-
-.ql-formats {
-  border: none;
-  display: flex;
-  margin-right: 1% !important;
-  padding: 0.1% !important;
-  gap: 0 !important;
-}
-
-.ql-toolbar-right {
-  position: absolute;
-  right: 10px;
-}
-
-.ql-active {
-  color: #474d66 !important;
-}
-
-.ql-active > img {
-  background-color: skyblue !important;
-  border-radius: 5px;
-}
-
-.ql-editor {
-  width: 100%;
-  border: none;
-  min-height: 480px;
-  height: auto;
-  resize: none;
-  padding: 15px;
-  margin-top: 10px;
-  box-sizing: border-box;
-  line-height: 2;
-}
-
-.ql-editor.ql-blank::before {
-  font-family: 'Pretendard';
-  content: attr(data-placeholder);
-  font-size: 20px;
-  font-weight: 400;
-  font-style: normal;
-  color: #c6cada;
-  top: 7px;
-}
-
-@media (max-width: 768px) {
-  .buttons {
-    display: none;
-  }
-  .userName {
-    display: none;
-  }
-}
-
-@media (max-width: 480px) {
-  .ql-formats {
-    margin-right: 3px !important;
-    padding: 4px !important;
-  }
-  .ql-container {
-    min-height: 280px;
-  }
-  .ql-toolbar .ql-formats button {
-    width: 20px;
-    height: 20px;
-    padding: 0;
-  }
-}
